requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
opencv-python==4.8.1.78
numpy==1.26.2
python-multipart==0.0.6
